+++ tests/src/rules/extensions.js
@@ -293,6 +293,7 @@ ruleTester.run('extensions', rule, {
         import bar from './bar.json'
         import Component from './Component'
         import baz from 'foo/baz'
+        import baw from '@scoped/baw/import'
         import express from 'express'
       `,
       options: [ 'always', {ignorePackages: true} ],
@@ -301,10 +302,6 @@ ruleTester.run('extensions', rule, {
           message: 'Missing file extension for "./Component"',
           line: 4,
           column: 31,
-        }, {
-          message: 'Missing file extension for "foo/baz"',
-          line: 5,
-          column: 25,
         },
       ],
     }),
@@ -315,6 +312,7 @@ ruleTester.run('extensions', rule, {
         import bar from './bar.json'
         import Component from './Component'
         import baz from 'foo/baz'
+        import baw from '@scoped/baw/import'
         import express from 'express'
       `,
       options: [ 'ignorePackages' ],
@@ -323,10 +321,6 @@ ruleTester.run('extensions', rule, {
           message: 'Missing file extension for "./Component"',
           line: 4,
           column: 31,
-        }, {
-          message: 'Missing file extension for "foo/baz"',
-          line: 5,
-          column: 25,
         },
       ],
     }),